pyproject.toml
[project]
name = "todo-api"
version = "1.0.0"
description = "Micro-API de Gerenciamento de Tarefas (MVP)"
authors = [
    {name = "Danilo Bressane"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "sqlalchemy>=2.0.29",
    "aiosqlite>=0.20.0",
    "pydantic>=2.6.4",
    "pydantic-settings>=2.2.1",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",
    "greenlet>=3.0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "respx>=0.21.1",
    "coverage>=7.4.1",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
